rename bodyStream to body_stream as expected
