+++ sdk/core/azure-core/inc/http/body_stream.hpp
