+++ sdk/core/azure-core/inc/http/http.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "bodyStream.hpp"
+#include "body_stream.hpp"
 
 #include <algorithm>
 #include <internal/contract.hpp>
