+++ sdk/core/azure-core/src/http/body_stream.cpp
@@ -16,7 +16,7 @@
 #include <cstdint>
 #include <cstdio>
 #include <cstring>
-#include <http/bodyStream.hpp>
+#include <http/body_stream.hpp>
 #include <memory>
 #include <vector>
